+++ packages/simulator-renderer/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @harrywan/letgo-simulator-renderer
 
+## 0.0.0-beta.21
+
+### Patch Changes
+
+- Updated dependencies
+  - @harrywan/letgo-common@0.0.0-beta.21
+  - @harrywan/letgo-components@0.0.0-beta.21
+  - @harrywan/letgo-designer@0.0.0-beta.21
+  - @harrywan/letgo-renderer@0.0.0-beta.21
+
 ## 0.0.0-beta.20
 
 ### Patch Changes